# Switch to different implementation of xxhash