--- a/.github/workflows/go.yml
+++ b/.github/workflows/go.yml
@@ -7,11 +7,14 @@ on:
 jobs:
   test:
     runs-on: ubuntu-latest
+    strategy:
+      matrix:
+        go: ['1.15', '1.14']
     steps:
-    - name: Install Go
-      uses: actions/setup-go@v1
+    - name: Install Go ${{ matrix.go }}
+      uses: actions/setup-go@v2
       with:
-        go-version: 1.14.x
+        go-version: ${{ matrix.go }}
     - name: Checkout code
       uses: actions/checkout@v2
     - name: Build and run tests