--- a/go.mod
+++ b/go.mod
@@ -1,8 +1,13 @@
 module github.com/eiri/dieci
 
-go 1.14
+go 1.15
 
 require (
-	github.com/cespare/xxhash v1.1.0
-	github.com/stretchr/testify v1.5.1
+	github.com/OneOfOne/xxhash v1.2.8
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/kr/text v0.2.0 // indirect
+	github.com/niemeyer/pretty v0.0.0-20200227124842-a10e7caefd8e // indirect
+	github.com/stretchr/testify v1.6.1
+	gopkg.in/check.v1 v1.0.0-20200902074654-038fdea0a05b // indirect
+	gopkg.in/yaml.v3 v3.0.0-20200615113413-eeeca48fe776 // indirect
 )