It was an indirect dep anyway and it seems to be ~same on bench

```
# run 1
benchcmp before.bench after.bench
benchmark                old ns/op     new ns/op     delta
BenchmarkOpenIndex-4     74993851      72243543      -3.67%
BenchmarkOpen-4          62336029      60922477      -2.27%
BenchmarkWrite-4         12056         12010         -0.38%
BenchmarkRead-4          1157          1144          -1.12%

benchmark                old allocs     new allocs     delta
BenchmarkOpenIndex-4     9701           9675           -0.27%
BenchmarkOpen-4          9689           9696           +0.07%
BenchmarkWrite-4         1              1              +0.00%
BenchmarkRead-4          1              1              +0.00%

benchmark                old bytes     new bytes     delta
BenchmarkOpenIndex-4     31402580      31401187      -0.00%
BenchmarkOpen-4          31398387      31406632      +0.03%
BenchmarkWrite-4         1244          1244          +0.00%
BenchmarkRead-4          32            32            +0.00%

# run 2
benchcmp before.bench after.bench
benchmark                old ns/op     new ns/op     delta
BenchmarkOpenIndex-4     71048252      72197142      +1.62%
BenchmarkOpen-4          60329998      60784354      +0.75%
BenchmarkWrite-4         13224         13756         +4.02%
BenchmarkRead-4          1182          1139          -3.64%

benchmark                old allocs     new allocs     delta
BenchmarkOpenIndex-4     9663           9688           +0.26%
BenchmarkOpen-4          9658           9672           +0.14%
BenchmarkWrite-4         1              1              +0.00%
BenchmarkRead-4          1              1              +0.00%

benchmark                old bytes     new bytes     delta
BenchmarkOpenIndex-4     31396797      31403032      +0.02%
BenchmarkOpen-4          31392829      31400654      +0.02%
BenchmarkWrite-4         1233          1247          +1.14%
BenchmarkRead-4          32            32            +0.00%
``` 

Also update deps and make action to check on golang 1.14 and 1.15